requirements.txt
langchain
openai
clickhouse-connect
clickhouse-sqlalchemy
python-dotenv